% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annual-to-bey.R
\name{annual_to_bey}
\alias{annual_to_bey}
\title{Convert the annual rate to the bond equivalent yield}
\usage{
annual_to_bey(annual_rate, settlement_date, maturity_date)
}
\arguments{
\item{annual_rate}{Annual rate.}

\item{settlement_date}{Settlement date.}

\item{maturity_date}{Maturity date.}
}
\value{
bey Bond equivalent yield.
}
\description{
Convert the annual rate to the bond equivalent yield
}
\examples{
settlement_date <- lubridate::ymd( "2015-01-30" )
maturity_date <- lubridate::ymd( "2015-01-30" )
annual_to_bey( annual_rate = 0.02, settlement_date, maturity_date )
}

