% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bey-indemnity.R
\name{bey_indemnity}
\alias{bey_indemnity}
\title{Calculate the interpolated bond equivalent yield for indemnity calculation.}
\usage{
bey_indemnity(bey_short, bey_long, date_short, date_long, date_wal)
}
\arguments{
\item{bey_short}{Bond equivalent yield of short-maturity tenor.}

\item{bey_long}{Bond equivalent yield of long-maturity tenor.}

\item{date_short}{Maturity date of short-maturity tenor.}

\item{date_long}{Maturity date of long-maturity tenor.}

\item{date_wal}{Weighted average life as a date.}
}
\value{
bey_interpolated Interpolated yield for indemnity calculation.
}
\description{
Link for more details is here:
\url{https://www.cmhc-schl.gc.ca/en/hoficlincl/mobase/upload/nha_mbs_indemnity_calculation_methodology.pdf}
}
\examples{
bey_short <- 1.363
bey_long <- 1.501
date_short <- ymd( "2016-06-01" )
date_long <- ymd( "2017-09-01" )
date_wal <- ymd( "2016-11-23" )
bey_indemnity( bey_short, bey_long, date_short, date_long, date_wal )
}

