% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interpolate-bey.R
\name{interpolate_bey}
\alias{interpolate_bey}
\title{Find the interpolated bond equivalent yield for indemnity calculation.}
\usage{
interpolate_bey(issue_date, settlement_date, maturity_date, maturity_dt,
  yields_dt, wal, cmhc_audit = TRUE)
}
\arguments{
\item{issue_date}{Issue date.}

\item{settlement_date}{Settlement date.}

\item{maturity_date}{Maturity date.}

\item{yields_dt}{\code{Datatable} of historical yield curves.}

\item{wal}{Weighted average life of security (NHA MBS pool).}
}
\value{
bey_interpolated Interpolated BEY for indemnity calculation.
}
\description{
Find the interpolated bond equivalent yield for indemnity calculation.
}
\examples{
issue_date <- lubridate::ymd( "2015-01-30" )
reporting_date <- lubridate::ymd( "2015-01-30" )
data( bloomberg_goc, package = "blpxl" )
data( bloomberg_cad_maturity, package = "blpxl" )
interpolate_bey( issue_date, reporting_date, wal, yields_dt = bloomberg_goc, maturity_dt = bloomberg_cad_maturity )
}

