% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interpolation-pairs.R
\name{interpolation_pairs}
\alias{interpolation_pairs}
\title{Get interpolation pairs for the yields, maturity dates and tenors.}
\usage{
interpolation_pairs(retrieval_date, wal_date, maturity_dt, yields_dt)
}
\arguments{
\item{retrieval_date}{Retrieval (anchor) date.}

\item{wal_date}{Weighted average life as a date.}

\item{maturity_dt}{\code{Datatable} of maturity of each (CAD curve) tenor.}

\item{yields_dt}{\code{Datatable} of yield of each (CAD curve) tenor.}
}
\value{
yields Yields of short and long tenor to interpolate between.
}
\description{
Get interpolation pairs for the yields, maturity dates and tenors.
}
\examples{
retrieval_date <- lubridate::ymd( "2013-03-01" )
wal_date <- lubridate::ymd( "2014-06-21" )
data( bloomberg_goc, package = "blpxl" )
data( bloomberg_cad_maturity, package = "blpxl" )
list_args_pt1 <- list( wal_date = wal_date, retrieval_date = retrieval_date )
list_args_pt2 <- list( maturity_dt = bloomberg_cad_maturity, yields_dt = bloomberg_goc )
do.call( interpolation_pairs, args = c( list_args_pt1, list_args_pt2 ) )
}

