% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/must-convert-to-bey.R
\name{must_convert_to_bey}
\alias{must_convert_to_bey}
\title{Check if tenor must be converted to the bond equivalent yield.}
\usage{
must_convert_to_bey(tenor, tenors_to_convert = c("OVERNIGHT", "MONTH3",
  "MONTH6", "MONTH12"))
}
\arguments{
\item{tenor}{Tenor (or pillar) of the security.}

\item{tenors_to_convert}{Character vector of tenors to convert.}
}
\value{
\code{TRUE} if tenor must be converted to bond equivalent yield.
}
\description{
Check if tenor must be converted to the bond equivalent yield.
}
\examples{
must_convert_to_bey( tenor = "MONTH3" )
must_convert_to_bey( tenor = "MONTH24" )
must_convert_to_bey( tenor = "OVERNIGHT" )
}

