% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wal-to-date.R
\name{wal_to_date}
\alias{wal_to_date}
\title{Convert weighted average life in years (decimal) to date}
\usage{
wal_to_date(settlement_date, wal)
}
\arguments{
\item{settlement_date}{Settlement date.}

\item{wal}{Weighted average life of NHA MBS pool.}
}
\value{
wal_date Weighted average life as a date.
}
\description{
Link for more details is here:
\url{https://www.cmhc-schl.gc.ca/en/hoficlincl/mobase/upload/nha_mbs_indemnity_calculation_methodology.pdf}
}
\examples{
settlement_date <- lubridate::ymd( "2015-01-30" )
wal_to_date( settlement_date, 3.25 )
wal_to_date( settlement_date, 4.89 )
}

