% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ycurve-for-interpolation.R
\name{ycurve_for_interpolation}
\alias{ycurve_for_interpolation}
\title{Get the yield curve to interpolate for indemnity calculation}
\usage{
ycurve_for_interpolation(retrieval_date, ycurve_dt, date_re = "date")
}
\arguments{
\item{retrieval_date}{Retrieval (anchor) date.}

\item{ycurve_dt}{Datatable of historical yield curves.}

\item{date_re}{Regex to uniquely match the column with dates in \code{yields_dt}.}
}
\value{
dt   Data.table of yield curve for indemnity calculation
}
\description{
Get the yield curve to interpolate for indemnity calculation
}
\examples{
retrieval_date <- lubridate::ymd( "2013-03-01" )
data( bloomberg_goc, package = "blpxl" )
ycurve_for_interpolation( retrieval_date, ycurve_dt = bloomberg_goc )
}

