% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ycurve-retrieval-date.R
\name{ycurve_retrieval_date}
\alias{ycurve_retrieval_date}
\title{Get the retrieval date for the yield curve for indemnity interpolation.}
\usage{
ycurve_retrieval_date(issue_date, settlement_date, cmhc_audit = TRUE)
}
\arguments{
\item{issue_date}{Issue date.}

\item{settlement_date}{Settlement date.}

\item{cmhc_audit}{\code{TRUE}. Applies CMHC Audit: first business day of the month.}
}
\value{
retrieval_date Date on which the yield curve is interpolated.
}
\description{
Get the retrieval date for the yield curve for indemnity interpolation.
}
\examples{
issue_date <- lubridate::ymd( "2015-01-30" )
settlement_date <- lubridate::ymd( "2015-01-30" )
ycurve_retrieval_date( issue_date, settlement_date )
}

